microservice/requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
feature_engine>=1.6.0
fastapi
uvicorn
pyyaml
pydantic